Ревью. Сохранение input keyWord и переключателя в localStorage in proc
